svm-conformance: move serialization harness
